+++ Sources/MapboxNavigation/Constants.swift
@@ -110,8 +110,3 @@ public typealias CarPlayUserInfo = [String: Any?]
  to a color defined in `InstructionsCardContainerView.highlightedBackgroundColor`.
  */
 let InstructionCardHighlightDistance: CLLocationDistance = 152.4 // 500 ft
-
-/**
- Default tolerance value used for configuring the underlying map source of route line, maneuver arrow and restricted areas.
- */
-let RouteSourcesTolerance: Double = 0.375